move: CASSETTE_DIR makro and scrub_body function to init
